encapsulate converting in a service class
